+++ actors/src/test/kotlin/ktx/actors/EventsTest.kt
@@ -83,6 +83,46 @@ class EventsTest {
     assertTrue(listener in actor.listeners)
   }
 
+  @Test
+  fun `should attach ClickListener triggered on enter events`() {
+    val actor = Actor()
+
+    val listener = actor.onEnter { }
+
+    assertNotNull(listener)
+    assertTrue(listener in actor.listeners)
+  }
+
+  @Test
+  fun `should attach InputListener triggered on exit events`() {
+    val actor = Actor()
+
+    val listener = actor.onExit { }
+
+    assertNotNull(listener)
+    assertTrue(listener in actor.listeners)
+  }
+
+  @Test
+  fun `should attach ClickListener consuming onEnter events with local coordinates`() {
+    val actor = Actor()
+
+    val listener = actor.onEnterEvent { event, x, y -> }
+
+    assertNotNull(listener)
+    assertTrue(listener in actor.listeners)
+  }
+
+  @Test
+  fun `should attach ClickListener consuming onExit events with local coordinates`() {
+    val actor = Actor()
+
+    val listener = actor.onExitEvent { event, x, y -> }
+
+    assertNotNull(listener)
+    assertTrue(listener in actor.listeners)
+  }
+
   class SampleNode : Tree.Node<SampleNode, Any, Actor>(mock())
 
   @Test